requirements.txt
dash_bootstrap_components==1.6.0
dash==2.18.2
plotly==5.24.1
OpenAI==0.28.0
langchain==0.0.284
python-dotenv==1.0.0
unstructured==0.9.2
tiktoken==0.4.0
faiss-cpu==1.7.4
libmagic==1.0
python-magic==0.4.27
ython-magic-bin==0.4.14
OpenAI == 0.28.0
